added more unit tests and fixed some issues in the chesslib to make the tests pass (still WIP)

+++ ChessLib/include/chessposition.h
@@ -12,7 +12,7 @@
              D E F I N E    F U N C T I O N S
    ==================================================== */
 
-ChessPosition create_position(int row, int column);
+ChessPosition create_position(int8_t row, int8_t column);
 
 int8_t get_row(ChessPosition position);
 int8_t get_column(ChessPosition position);
